moved external api call to a new folder
