+++ app/src/services/external/exchangeRate.js
@@ -1,5 +1,5 @@
 const axios = require('axios');
-import { env } from '../config';
+import { env } from '../../config';
 
 async function getRate(from, to, reference) {
     const response = await axios.get(`${env.api.url}?fsym=${reference}&tsyms=${from},${to}`);
